Remove custom useWeather hook
